+++ packages/sqlite/src/SqlitePlatform.ts
@@ -102,21 +102,26 @@ export class SqlitePlatform extends AbstractSqlPlatform {
     return escape(value, true, this.timezone);
   }
 
-  getSearchJsonPropertyKey(path: string[], type: string): string {
+  getSearchJsonPropertyKey(path: string[], type: string, aliased: boolean): string {
     const [a, ...b] = path;
-    return expr(alias => `json_extract(${this.quoteIdentifier(`${alias}.${a}`)}, '$.${b.join('.')}')`);
-  }
 
-  getDefaultIntegrityRule(): string {
-    return 'no action';
+    if (aliased) {
+      return expr(alias => `json_extract(${this.quoteIdentifier(`${alias}.${a}`)}, '$.${b.join('.')}')`);
+    }
+
+    return `json_extract(${this.quoteIdentifier(a)}, '$.${b.join('.')}')`;
   }
 
   getIndexName(tableName: string, columns: string[], type: 'index' | 'unique' | 'foreign' | 'primary' | 'sequence'): string {
     if (type === 'primary') {
-      return 'primary';
+      return this.getDefaultPrimaryName(tableName, columns);
     }
 
     return super.getIndexName(tableName, columns, type);
+  }
+
+  getDefaultPrimaryName(tableName: string, columns: string[]): string {
+    return 'primary';
   }
 
   supportsDownMigrations(): boolean {