+++ tests/EntityManager.postgre.test.ts
@@ -1,9 +1,9 @@
 import { v4 } from 'uuid';
-import type { EventSubscriber, ChangeSet, AnyEntity, FlushEventArgs } from '@mikro-orm/core';
+import type { EventSubscriber, ChangeSet, AnyEntity, FlushEventArgs, FilterQuery, EntityDTO } from '@mikro-orm/core';
 import {
   Collection, Configuration, EntityManager, LockMode, MikroORM, QueryFlag, QueryOrder, Reference, ValidationError, ChangeSetType, wrap, expr,
   UniqueConstraintViolationException, TableNotFoundException, NotNullConstraintViolationException, TableExistsException, SyntaxErrorException,
-  NonUniqueFieldNameException, InvalidFieldNameException, LoadStrategy, IsolationLevel, PopulateHint,
+  NonUniqueFieldNameException, InvalidFieldNameException, LoadStrategy, IsolationLevel, PopulateHint, FlushMode,
 } from '@mikro-orm/core';
 import { PostgreSqlDriver, PostgreSqlConnection } from '@mikro-orm/postgresql';
 import { Address2, Author2, Book2, BookTag2, FooBar2, FooBaz2, Publisher2, PublisherType, PublisherType2, Test2, Label2 } from './entities-sql';
@@ -315,6 +315,11 @@ describe('EntityManagerPostgre', () => {
     expect(mock.mock.calls[5][0]).toMatch(`release savepoint trx`);
     expect(mock.mock.calls[6][0]).toMatch(`select "b0"."uuid_pk", "b0"."created_at", "b0"."title", "b0"."price", "b0"."double", "b0"."meta", "b0"."author_id", "b0"."publisher_id", "b0".price * 1.19 as "price_taxed" from "book2" as "b0" where "b0"."author_id" is not null and "b0"."publisher_id" = $1 order by "b0"."uuid_pk" asc for update`);
     expect(mock.mock.calls[7][0]).toMatch(`commit`);
+  });
+
+  test('em.commit/rollback validation', async () => {
+    await expect(orm.em.commit()).rejects.toThrowError('An open transaction is required for this operation');
+    await expect(orm.em.rollback()).rejects.toThrowError('An open transaction is required for this operation');
   });
 
   test('should load entities', async () => {
@@ -1610,11 +1615,12 @@ describe('EntityManagerPostgre', () => {
     await orm.em.flush();
     orm.em.clear();
 
-    // without paginate flag it fails to get 5 records
+    // without paginate flag it fails to get only 2 records (we need to explicitly disable it)
     const res1 = await orm.em.find(Author2, { books: { title: /^Bible/ } }, {
       orderBy: { name: QueryOrder.ASC, books: { title: QueryOrder.ASC } },
       offset: 3,
       limit: 5,
+      flags: [QueryFlag.DISABLE_PAGINATE],
     });
 
     expect(res1).toHaveLength(2);
@@ -2029,6 +2035,31 @@ describe('EntityManagerPostgre', () => {
     expect(mock.mock.calls[11][0]).toMatch(`commit`);
 
     mock.mockRestore();
+  });
+
+  test('GH #2934', async () => {
+    const users = [
+      { name: 'A', email: 'A' },
+      { name: 'B', email: 'B' },
+      { name: 'C', email: 'C' },
+      { name: 'D', email: 'D' },
+    ];
+
+    async function saveUser(options: FilterQuery<Author2>): Promise<Author2> {
+      let user = await orm.em.findOne(Author2, options);
+
+      if (!user) {
+        user = orm.em.create(Author2, options as any);
+        await orm.em.persistAndFlush(user);
+      }
+
+      expect(user.id).toBeDefined();
+
+      return user;
+    }
+
+    const res = await Promise.all(users.map(userData => saveUser(userData)));
+    res.forEach(user => expect(user.id).toBeDefined());
   });
 
   test('required fields validation', async () => {