+++ tests/issues/GH1224.test.ts
@@ -26,7 +26,7 @@ class B {
 class A {
 
   [PrimaryKeyType]?: number;
-  [PrimaryKeyProp]: 'node';
+  [PrimaryKeyProp]?: 'node';
   @OneToOne({ entity: 'Node', wrappedReference: true, primary: true, onDelete: 'cascade', onUpdateIntegrity: 'cascade' })
   node!: IdentifiedReference<Node>;